refactor: Refactored the plugin test bases so that GrpcPluginTestBase extends PluginTestBase
